Turn visible menu items having url="#" without linkOptions=>submit
This fix allows a more "intuitive" use of this extension.
Sometimes we need specify some item in the menu just for group subitems
so that we use url="#" on it without any submit url in linkOptions.
This fix turn that items visible for default. We no need more to force
a visible=true;